Organize BarChart files
